[REF] hr_payroll_document: In tests, do not repeat wizard creation
